backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.3.0
python-decouple==3.8
psycopg2-binary>=2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-admin-interface>=0.28.0
django-colorfield>=0.11.0
djangorestframework-simplejwt==5.3.1
django-filter>=24.1
